+++ tests/testthat/test-lancelot-support.R
@@ -420,6 +420,9 @@ test_that("lancelot_index identifies ICU and D_tot in real model", {
   expect_equal(
     names(index$run),
     c("time", "icu", "general", "deaths_carehomes_inc", "deaths_comm_inc",
+      "deaths_comm_0_49_inc", "deaths_comm_50_54_inc", "deaths_comm_55_59_inc",
+      "deaths_comm_60_64_inc", "deaths_comm_65_69_inc", "deaths_comm_70_74_inc",
+      "deaths_comm_75_79_inc", "deaths_comm_80_plus_inc",
       "deaths_hosp_inc", "deaths_hosp_0_49_inc", "deaths_hosp_50_54_inc",
       "deaths_hosp_55_59_inc", "deaths_hosp_60_64_inc", "deaths_hosp_65_69_inc",
       "deaths_hosp_70_74_inc", "deaths_hosp_75_79_inc",
@@ -539,6 +542,14 @@ test_that("lancelot_compare combines likelihood correctly", {
     general = 20:25,
     deaths_carehomes_inc = 2:7,
     deaths_comm_inc = 1:6,
+    deaths_comm_0_49_inc = 1:6,
+    deaths_comm_50_54_inc = 1:6,
+    deaths_comm_55_59_inc = 2:7,
+    deaths_comm_60_64_inc = 2:7,
+    deaths_comm_65_69_inc = 3:8,
+    deaths_comm_70_74_inc = 3:8,
+    deaths_comm_75_79_inc = 4:9,
+    deaths_comm_80_plus_inc = 4:9,
     deaths_hosp_inc = 3:8,
     deaths_hosp_0_49_inc = 1:6,
     deaths_hosp_50_54_inc = 1:6,
@@ -587,6 +598,14 @@ test_that("lancelot_compare combines likelihood correctly", {
     deaths_hosp_75_79 = 7,
     deaths_hosp_80_plus = 8,
     deaths_comm = 3,
+    deaths_comm_0_49 = 1,
+    deaths_comm_50_54 = 2,
+    deaths_comm_55_59 = 3,
+    deaths_comm_60_64 = 4,
+    deaths_comm_65_69 = 5,
+    deaths_comm_70_74 = 6,
+    deaths_comm_75_79 = 7,
+    deaths_comm_80_plus = 8,
     deaths = 8,
     deaths_non_hosp = 6,
     admitted = 53,
@@ -887,19 +906,30 @@ test_that("lancelot_check_data requires consistent deaths", {
 })
 
 
-test_that("lancelot_check_data disallows double fitting to deaths", {
+test_that("lancelot_check_data disallows double fitting to hospital deaths", {
   ## Specifically, disallow double fitting to aggregated and
   ## disaggregated hospital deaths
   data <- lancelot_simple_data(read_csv(sircovid_file("extdata/example.csv")))
+
+  data$deaths <- NA
+  data$deaths_hosp <- 10
+
+  expect_silent(lancelot_check_data(data))
+
+  ## Allow age disaggregated and aggregated on different days
+  data$deaths_hosp[11:20] <- NA
+  data$deaths_hosp_60_64[11:20] <- 3
+  data$deaths_hosp_80_plus[11:20] <- 7
 
   expect_silent(lancelot_check_data(data))
 
   ## Throw error
   data$deaths_hosp_0_49 <- 1
-  data_deaths_hosp_65_69 <- 4
+  data$deaths_hosp_65_69 <- 4
   expect_error(
     lancelot_check_data(data),
-    "Cannot fit to all ages aggregated for deaths if fitting to any sub-groups")
+    paste0("Cannot fit to all ages aggregated for hospital deaths if fitting ",
+           "to any sub-groups"))
 
   ## Add populations and throw regional error
   data2 <- cbind(
@@ -908,7 +938,43 @@ test_that("lancelot_check_data disallows double fitting to deaths", {
 
   expect_error(
     lancelot_check_data(data2),
-    "london: Cannot fit to all ages aggregated for deaths if")
+    "london: Cannot fit to all ages aggregated for hospital deaths if")
+})
+
+
+test_that("lancelot_check_data disallows double fitting to community deaths", {
+  ## Specifically, disallow double fitting to aggregated and
+  ## disaggregated community deaths
+  data <- lancelot_simple_data(read_csv(sircovid_file("extdata/example.csv")))
+
+  data$deaths <- NA
+  data$deaths_comm <- 10
+
+  expect_silent(lancelot_check_data(data))
+
+  ## Allow age disaggregated and aggregated on different days
+  data$deaths_comm[11:20] <- NA
+  data$deaths_comm_60_64[11:20] <- 3
+  data$deaths_comm_80_plus[11:20] <- 7
+
+  expect_silent(lancelot_check_data(data))
+
+  ## Throw error
+  data$deaths_comm_0_49 <- 1
+  data$deaths_comm_65_69 <- 4
+  expect_error(
+    lancelot_check_data(data),
+    paste0("Cannot fit to all ages aggregated for community deaths if fitting ",
+           "to any sub-groups"))
+
+  ## Add populations and throw regional error
+  data2 <- cbind(
+    rbind(data, data),
+    region = factor(rep(c("london", "south_east"), each = nrow(data))))
+
+  expect_error(
+    lancelot_check_data(data2),
+    "london: Cannot fit to all ages aggregated for community deaths if")
 })